refactor: extract application route logic into controller and validator
